requirements.txt
fastapi>=0.110.0
starlette>=0.36.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pydantic>=2.0.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
jinja2>=3.1.0
python-multipart>=0.0.9
itsdangerous>=2.1.0
psycopg2-binary>=2.9.0
